reworking error handling in the api
This is still in progress, im going to rename ApiError to PublicApiError and have that only contain information relevant to clients to know, internal errors will be logged and handled directly rather than trying to make a giant enum fitting them all.

Offcourses stores and the like gets a error enum as they are delegating the error handling to the caller.

But the "end user" is the api handlers and we should not try to return a general "ApiError" other than for the response purpose.